backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
email-validator==2.2.0
reportlab==4.2.5
boto3==1.35.3
